requirements.txt
playwright==1.52.0
python-telegram-bot==21.11.1
